exceptions do controller done

+++ src/main/java/com/tabd/mcdonalds/service/McDonaldsService.java
@@ -17,24 +17,65 @@ public class McDonaldsService {
     }
 
     public void save(McDonalds mcDonalds){
-        this.mcDonaldsRepository.save(mcDonalds);
+        if(mcDonaldsRepository.findById(mcDonalds.getN_Pedido()).isPresent()){
+            throw new PedidoRepeatedException(mcDonalds.getN_Pedido());
+        }
+        if(verifyFields(mcDonalds)){
+            this.mcDonaldsRepository.save(mcDonalds);
+        }
     }
 
     public void delete(Integer n_Pedido){
-        this.mcDonaldsRepository.deleteById(n_Pedido);
+        if(mcDonaldsRepository.findById(n_Pedido).isPresent()){
+            this.mcDonaldsRepository.deleteById(n_Pedido);
+        }
+        else throw new PedidoNotFoundException(n_Pedido);
     }
 
-    public Optional<McDonalds> findById(Integer n_Pedido){
-        return this.mcDonaldsRepository.findById(n_Pedido);
+    public McDonalds findById(Integer n_Pedido){
+        return mcDonaldsRepository.findById(n_Pedido)
+                .orElseThrow(() -> new PedidoNotFoundException(n_Pedido));
     }
 
     public List<McDonalds> list() {
-        return this.mcDonaldsRepository.findAll();
+        List<McDonalds> lista = this.mcDonaldsRepository.findAll();
+        if (lista.isEmpty()){
+            throw( new PedidoEmptyException());
+        }
+        else return lista;
     }
 
-    public void update(Integer n_Pedido, McDonalds mcDonaldsNovo){
-        Optional<McDonalds> mcDonaldsAntigo = this.mcDonaldsRepository.findById(n_Pedido);
-        mcDonaldsNovo.setN_Pedido(mcDonaldsAntigo.get().getN_Pedido());
-        this.mcDonaldsRepository.save(mcDonaldsNovo);
+    public void update(Integer n_Pedido, McDonalds mcDonaldsNew){
+        McDonalds mcDonaldsOld = this.findById(n_Pedido);
+        replaceFields(mcDonaldsNew, mcDonaldsOld);
+        this.mcDonaldsRepository.save(mcDonaldsNew);
+    }
+
+    private boolean verifyFields(McDonalds mcDonalds){
+        if(mcDonalds.getN_Pedido() == 0
+                || mcDonalds.getItem() == null
+                || mcDonalds.getQtd() == 0
+                || mcDonalds.getValor() == 0.0
+        ){
+            throw new PedidoMissingValueException();
+        }
+        return true;
+    }
+
+    private McDonalds replaceFields(McDonalds mcDonaldsNew, McDonalds mcDonaldsOld){
+        mcDonaldsNew.setN_Pedido(mcDonaldsOld.getN_Pedido());
+
+        if(mcDonaldsNew.getItem() == null){
+            mcDonaldsNew.setItem(mcDonaldsOld.getItem());
+        }
+
+        if(mcDonaldsNew.getValor() == 0.0){
+            mcDonaldsNew.setValor(mcDonaldsOld.getValor());
+        }
+
+        if(mcDonaldsNew.getQtd() == 0){
+            mcDonaldsNew.setQtd(mcDonaldsOld.getQtd());
+        }
+        return mcDonaldsNew;
     }
 }
